// ui/src/generated/MidiMapCommand.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { MidiMapMode } from "./MidiMapMode";
import type { MidiMapTarget } from "./MidiMapTarget";
import type { MidiMapping } from "./MidiMapping";
import type { MidiMappingId } from "./MidiMappingId";

export type MidiMapCommand = { "type": "Learn", target: MidiMapTarget | null, } | { "type": "Map", mapping: MidiMapping, } | { "type": "Edit", id: MidiMappingId, min: number | null, max: number | null, mode: MidiMapMode | null, } | { "type": "Unmap", ids: Array<MidiMappingId>, } | { "type": "List" };